Increase compression further by removing duplicate rows
The MSP uses the "row_offset" array to discover where to find the
beginning of the array in memory.  This patch makes img2c.py look to
see if we've already output a row that's identical to the one we're
currently examining.  If we have, then it simply references that one
in the row_offset array instead of outputting new data.